backend/pyproject.toml
[project]
name = "screen-ai-backend"
version = "0.1.0"
description = "Backend API for Screen AI drug discovery platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.2",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]
